feat: convert controller.js to controller.ts
